fix symbol request (#6)

+++ backend/Naninovel.Language.Test/SymbolTest.cs
@@ -47,7 +47,7 @@ public class SymbolTest
     {
         var symbol = GetSymbol("");
         Assert.Equal("GenericTextLine", symbol.Name);
-        Assert.Equal(SymbolKind.String, symbol.Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Kind);
         Assert.Equal(Range.Empty, symbol.Range);
         Assert.Null(symbol.Detail);
         Assert.Null(symbol.Tags);
@@ -58,11 +58,11 @@ public class SymbolTest
     {
         var symbol = GetSymbol("# label");
         Assert.Equal("LabelLine", symbol.Name);
-        Assert.Equal(SymbolKind.Namespace, symbol.Kind);
+        Assert.Equal((int)SymbolKind.Namespace, symbol.Kind);
         Assert.Equal(new Range(new(0, 0), new(0, 7)), symbol.Range);
         Assert.Single(symbol.Children);
         Assert.Equal("LabelText", symbol.Children[0].Name);
-        Assert.Equal(SymbolKind.String, symbol.Children[0].Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Children[0].Kind);
         Assert.Equal(new Range(new(0, 2), new(0, 7)), symbol.Children[0].Range);
     }
 
@@ -71,11 +71,11 @@ public class SymbolTest
     {
         var symbol = GetSymbol("; comment");
         Assert.Equal("CommentLine", symbol.Name);
-        Assert.Equal(SymbolKind.String, symbol.Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Kind);
         Assert.Equal(new Range(new(0, 0), new(0, 9)), symbol.Range);
         Assert.Single(symbol.Children);
         Assert.Equal("CommentText", symbol.Children[0].Name);
-        Assert.Equal(SymbolKind.String, symbol.Children[0].Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Children[0].Kind);
         Assert.Equal(new Range(new(0, 2), new(0, 9)), symbol.Children[0].Range);
     }
 
@@ -84,39 +84,39 @@ public class SymbolTest
     {
         var symbol = GetSymbol("@cmd nameless param:{exp}x|#id|");
         Assert.Equal("CommandLine", symbol.Name);
-        Assert.Equal(SymbolKind.Struct, symbol.Kind);
+        Assert.Equal((int)SymbolKind.Struct, symbol.Kind);
         Assert.Equal(new Range(new(0, 0), new(0, 31)), symbol.Range);
         Assert.Single(symbol.Children);
         Assert.Equal("Command", symbol.Children[0].Name);
-        Assert.Equal(SymbolKind.Function, symbol.Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Function, symbol.Children[0].Kind);
         Assert.Equal(new Range(new(0, 1), new(0, 31)), symbol.Children[0].Range);
         Assert.Equal(3, symbol.Children[0].Children!.Count);
         Assert.Equal("CommandIdentifier", symbol.Children[0].Children[0].Name);
-        Assert.Equal(SymbolKind.Key, symbol.Children[0].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Key, symbol.Children[0].Children[0].Kind);
         Assert.Equal(new Range(new(0, 1), new(0, 4)), symbol.Children[0].Children[0].Range);
         Assert.Equal("Parameter", symbol.Children[0].Children[1].Name);
-        Assert.Equal(SymbolKind.Field, symbol.Children[0].Children[1].Kind);
+        Assert.Equal((int)SymbolKind.Field, symbol.Children[0].Children[1].Kind);
         Assert.Equal(new Range(new(0, 5), new(0, 13)), symbol.Children[0].Children[1].Range);
         Assert.Equal("Parameter", symbol.Children[0].Children[2].Name);
-        Assert.Equal(SymbolKind.Field, symbol.Children[0].Children[2].Kind);
+        Assert.Equal((int)SymbolKind.Field, symbol.Children[0].Children[2].Kind);
         Assert.Equal(new Range(new(0, 14), new(0, 31)), symbol.Children[0].Children[2].Range);
         Assert.Single(symbol.Children[0].Children[1].Children);
         Assert.Equal("ParameterValue", symbol.Children[0].Children[1].Children[0].Name);
-        Assert.Equal(SymbolKind.String, symbol.Children[0].Children[1].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Children[0].Children[1].Children[0].Kind);
         Assert.Equal(new Range(new(0, 5), new(0, 13)), symbol.Children[0].Children[1].Children[0].Range);
         Assert.Equal(2, symbol.Children[0].Children[2].Children!.Count);
         Assert.Equal("ParameterIdentifier", symbol.Children[0].Children[2].Children[0].Name);
-        Assert.Equal(SymbolKind.Key, symbol.Children[0].Children[2].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Key, symbol.Children[0].Children[2].Children[0].Kind);
         Assert.Equal(new Range(new(0, 14), new(0, 19)), symbol.Children[0].Children[2].Children[0].Range);
         Assert.Equal("ParameterValue", symbol.Children[0].Children[2].Children[1].Name);
-        Assert.Equal(SymbolKind.String, symbol.Children[0].Children[2].Children[1].Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Children[0].Children[2].Children[1].Kind);
         Assert.Equal(new Range(new(0, 20), new(0, 31)), symbol.Children[0].Children[2].Children[1].Range);
         Assert.Equal(2, symbol.Children[0].Children[2].Children[1].Children!.Count);
         Assert.Equal("Expression", symbol.Children[0].Children[2].Children[1].Children[0].Name);
-        Assert.Equal(SymbolKind.Property, symbol.Children[0].Children[2].Children[1].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Property, symbol.Children[0].Children[2].Children[1].Children[0].Kind);
         Assert.Equal(new Range(new(0, 20), new(0, 25)), symbol.Children[0].Children[2].Children[1].Children[0].Range);
         Assert.Equal("TextIdentifier", symbol.Children[0].Children[2].Children[1].Children[1].Name);
-        Assert.Equal(SymbolKind.String, symbol.Children[0].Children[2].Children[1].Children[1].Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Children[0].Children[2].Children[1].Children[1].Kind);
         Assert.Equal(new Range(new(0, 26), new(0, 31)), symbol.Children[0].Children[2].Children[1].Children[1].Range);
     }
 
@@ -125,39 +125,39 @@ public class SymbolTest
     {
         var symbol = GetSymbol("author.appearance: [cmd] text {exp} |#id|");
         Assert.Equal("GenericTextLine", symbol.Name);
-        Assert.Equal(SymbolKind.String, symbol.Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Kind);
         Assert.Equal(new Range(new(0, 0), new(0, 41)), symbol.Range);
         Assert.Equal(3, symbol.Children!.Count);
         Assert.Equal("GenericTextPrefix", symbol.Children[0].Name);
-        Assert.Equal(SymbolKind.Constant, symbol.Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Constant, symbol.Children[0].Kind);
         Assert.Equal(new Range(new(0, 0), new(0, 19)), symbol.Children[0].Range);
         Assert.Equal("InlinedCommand", symbol.Children[1].Name);
-        Assert.Equal(SymbolKind.Struct, symbol.Children[1].Kind);
+        Assert.Equal((int)SymbolKind.Struct, symbol.Children[1].Kind);
         Assert.Equal(new Range(new(0, 19), new(0, 24)), symbol.Children[1].Range);
         Assert.Equal("GenericText", symbol.Children[2].Name);
-        Assert.Equal(SymbolKind.String, symbol.Children[2].Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Children[2].Kind);
         Assert.Equal(new Range(new(0, 24), new(0, 41)), symbol.Children[2].Range);
         Assert.Equal(2, symbol.Children[2].Children!.Count);
         Assert.Equal("Expression", symbol.Children[2].Children[0].Name);
-        Assert.Equal(SymbolKind.Property, symbol.Children[2].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Property, symbol.Children[2].Children[0].Kind);
         Assert.Equal(new Range(new(0, 30), new(0, 35)), symbol.Children[2].Children[0].Range);
         Assert.Equal("TextIdentifier", symbol.Children[2].Children[1].Name);
-        Assert.Equal(SymbolKind.String, symbol.Children[2].Children[1].Kind);
+        Assert.Equal((int)SymbolKind.String, symbol.Children[2].Children[1].Kind);
         Assert.Equal(new Range(new(0, 36), new(0, 41)), symbol.Children[2].Children[1].Range);
         Assert.Equal(2, symbol.Children[0].Children!.Count);
         Assert.Equal("GenericTextAuthor", symbol.Children[0].Children[0].Name);
-        Assert.Equal(SymbolKind.Key, symbol.Children[0].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Key, symbol.Children[0].Children[0].Kind);
         Assert.Equal(new Range(new(0, 0), new(0, 6)), symbol.Children[0].Children[0].Range);
         Assert.Equal("GenericTextAuthorAppearance", symbol.Children[0].Children[1].Name);
-        Assert.Equal(SymbolKind.Enum, symbol.Children[0].Children[1].Kind);
+        Assert.Equal((int)SymbolKind.Enum, symbol.Children[0].Children[1].Kind);
         Assert.Equal(new Range(new(0, 7), new(0, 17)), symbol.Children[0].Children[1].Range);
         Assert.Single(symbol.Children[1].Children);
         Assert.Equal("Command", symbol.Children[1].Children[0].Name);
-        Assert.Equal(SymbolKind.Function, symbol.Children[1].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Function, symbol.Children[1].Children[0].Kind);
         Assert.Equal(new Range(new(0, 20), new(0, 23)), symbol.Children[1].Children[0].Range);
         Assert.Single(symbol.Children[1].Children[0].Children);
         Assert.Equal("CommandIdentifier", symbol.Children[1].Children[0].Children[0].Name);
-        Assert.Equal(SymbolKind.Key, symbol.Children[1].Children[0].Children[0].Kind);
+        Assert.Equal((int)SymbolKind.Key, symbol.Children[1].Children[0].Children[0].Kind);
         Assert.Equal(new Range(new(0, 20), new(0, 23)), symbol.Children[1].Children[0].Children[0].Range);
     }
 
@@ -194,13 +194,13 @@ public class SymbolTest
             new Parameter { Id = "named", ValueContainerType = ValueContainerType.NamedList }
         };
         meta.Commands = new[] { new Command { Id = "c", Parameters = parameters } };
-        Assert.Equal(SymbolKind.String, GetSymbol("@c str:x").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.Number, GetSymbol("@c int:x").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.Number, GetSymbol("@c dec:x").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.Boolean, GetSymbol("@c bool:true").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.Array, GetSymbol("@c list:,").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.Array, GetSymbol("@c named:.").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.String, GetSymbol("@c int:{}").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c str:x").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.Number, GetSymbol("@c int:x").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.Number, GetSymbol("@c dec:x").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.Boolean, GetSymbol("@c bool:true").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.Array, GetSymbol("@c list:,").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.Array, GetSymbol("@c named:.").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c int:{}").Children![0].Children![1].Children![1].Kind);
     }
 
     [Fact]
@@ -215,12 +215,12 @@ public class SymbolTest
             new Parameter { Id = "named", ValueContainerType = ValueContainerType.NamedList }
         };
         meta.Commands = new[] { new Command { Id = "c", Parameters = parameters } };
-        Assert.Equal(SymbolKind.String, GetSymbol("@c str:{}").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.String, GetSymbol("@c int:{}").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.String, GetSymbol("@c dec:{}").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.String, GetSymbol("@c bool:{}").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.String, GetSymbol("@c list:{}").Children![0].Children![1].Children![1].Kind);
-        Assert.Equal(SymbolKind.String, GetSymbol("@c named:{}").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c str:{}").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c int:{}").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c dec:{}").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c bool:{}").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c list:{}").Children![0].Children![1].Children![1].Kind);
+        Assert.Equal((int)SymbolKind.String, GetSymbol("@c named:{}").Children![0].Children![1].Children![1].Kind);
     }
 
     private IReadOnlyList<Symbol> GetSymbols (params string[] lines)
